feat: add reusable profile settings component
